+++ projects/igniteui-angular/src/lib/grids/grid-base.component.ts
@@ -86,6 +86,9 @@ import {
     IgxExcelStyleMovingTemplateDirective
 } from './filtering/excel-style/grid.excel-style-filtering.component';
 import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
+import { IgxGridFilteringRowComponent } from './filtering/grid-filtering-row.component';
+import { IgxDragIndicatorIconDirective } from './row-drag.directive';
+import { IgxDragDirective } from '../directives/dragdrop/dragdrop.directive';
 
 const MINIMUM_COLUMN_WIDTH = 136;
 const FILTER_ROW_HEIGHT = 50;
@@ -183,6 +186,13 @@ export interface IGridDataBindable {
     data: any[];
     filteredData: any[];
 }
+
+export interface IRowDragEndEventArgs {
+    owner: IgxDragDirective;
+    dragData: IgxRowComponent<IgxGridBaseComponent & IGridDataBindable>;
+}
+
+export interface IRowDragStartEventArgs extends IRowDragEndEventArgs, CancelableEventArgs { }
 
 export enum GridSummaryPosition {
     top = 'top',
@@ -516,9 +526,37 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
             // should selection persist?
             this.allRowsSelected = false;
             this.deselectAllRows();
-            this.markForCheck();
+            this.resetCachedWidths();
+            this.calculateGridSizes();
         }
     }
+
+    @Input()
+    get rowDraggable(): boolean {
+        return this._rowDrag;
+    }
+
+    /**
+     * Sets whether rows can be moved.
+     * ```html
+     * <igx-grid #grid [rowDraggable]="true"></igx-grid>
+     * ```
+	 * @memberof IgxGridBaseComponent
+     */
+    set rowDraggable(val: boolean) {
+        this._rowDrag = val;
+        if (this.gridAPI.grid) {
+            this.resetCachedWidths();
+            this.calculateGridSizes();
+        }
+    }
+
+    /**
+     * @hidden
+     * @internal
+     */
+    public rowDragging = false;
+
 
     /**
  * Sets whether the `IgxGridRowComponent` is editable.
@@ -684,7 +722,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     }
     public set columnWidth(value: string) {
         this._columnWidth = value;
-        this._columnWidthSetByUser = true;
+        this.columnWidthSetByUser = true;
     }
 
     /**
@@ -1444,6 +1482,20 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     @Output()
     public onFocusChange = new EventEmitter<IFocusChangeEventArgs>();
 
+     /**
+     * Emitted when start dragging a row.
+     * Return the dragged row.
+    */
+    @Output()
+    public onRowDragStart = new EventEmitter<IRowDragStartEventArgs>();
+
+    /**
+     * Emitted when dropping a row.
+     * Return the dropped row.
+    */
+    @Output()
+    public onRowDragEnd = new EventEmitter<IRowDragEndEventArgs>();
+
     /**
      * @hidden
      */
@@ -1479,6 +1531,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      */
     @ContentChild(IgxExcelStylePinningTemplateDirective, { read: IgxExcelStylePinningTemplateDirective })
     public excelStylePinningTemplateDirective: IgxExcelStylePinningTemplateDirective;
+
 
     /**
      * @hidden
@@ -1666,6 +1719,12 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     /**
      * @hidden
      */
+    @ViewChild('headerDragContainer')
+    public headerDragContainer: ElementRef;
+
+    /**
+     * @hidden
+     */
     @ViewChild('headerGroupContainer')
     public headerGroupContainer: ElementRef;
 
@@ -1674,6 +1733,12 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      */
     @ViewChild('headerCheckbox', { read: IgxCheckboxComponent })
     public headerCheckbox: IgxCheckboxComponent;
+
+    /**
+     * @hidden
+     */
+    @ViewChild('filteringRow', { read: IgxGridFilteringRowComponent })
+    public filteringRow: IgxGridFilteringRowComponent;
 
     /**
      * @hidden
@@ -1726,6 +1791,13 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     public get parentRowOutletDirective() {
         return null;
     }
+
+    /**
+     * @hidden
+     * @internal
+     */
+    @ViewChild('dragIndicatorIconBase', { read: TemplateRef })
+    public dragIndicatorIconBase: TemplateRef<any>;
 
     /**
      * @hidden
@@ -1916,7 +1988,9 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      */
     get maxLevelHeaderDepth() {
         if (this._maxLevelHeaderDepth === null) {
-            this._maxLevelHeaderDepth = this.columnList.reduce((acc, col) => Math.max(acc, col.level), 0);
+            this._maxLevelHeaderDepth = this.hasColumnLayouts ?
+                this.columnList.reduce((acc, col) => Math.max(acc, col.rowStart), 0) :
+                this.columnList.reduce((acc, col) => Math.max(acc, col.level), 0);
         }
         return this._maxLevelHeaderDepth;
     }
@@ -2270,10 +2344,6 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     /**
      * @hidden
      */
-    public calcRowCheckboxWidth = 0;
-    /**
-     * @hidden
-     */
     public calcHeight = 0;
     /**
      * @hidden
@@ -2314,6 +2384,11 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
         matchInfoCache: []
     };
 
+    /**
+     * @hidden
+     */
+    public columnWidthSetByUser = false;
+
     abstract data: any[];
     abstract filteredData: any[];
     // abstract dataLength;
@@ -2339,6 +2414,10 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      * @hidden
      */
     protected _rowSelection = false;
+    /**
+     * @hidden
+     */
+    protected _rowDrag = false;
     /**
      * @hidden
      */
@@ -2406,6 +2485,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     private _rowHeight;
     private _ngAfterViewInitPassed = false;
     private _horizontalForOfs;
+    private _multiRowLayoutRowSize = 1;
 
     // Caches
     private _totalWidth = NaN;
@@ -2417,7 +2497,6 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     private _columnGroups = false;
 
     private _columnWidth: string;
-    private _columnWidthSetByUser = false;
 
     private _defaultTargetRecordNumber = 10;
 
@@ -2557,13 +2636,11 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
             }
         });
 
-        this.onFilteringDone.pipe(destructor).subscribe(() => this.endEdit(true));
         this.onPagingDone.pipe(destructor).subscribe(() => {
             this.endEdit(true);
             this.selectionService.clear();
             this.selectionService.activeElement = null;
         });
-        this.onSortingDone.pipe(destructor).subscribe(() => this.endEdit(true));
 
         this.onColumnMoving.pipe(destructor).subscribe(() => this.endEdit(true));
         this.onColumnResized.pipe(destructor).subscribe(() => this.endEdit(true));
@@ -2799,12 +2876,16 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     /**
      * @hidden
      */
-    get headerCheckboxWidth() {
+    get headerFixedWidth() {
+        let width = 0;
         if (this.headerCheckboxContainer) {
-            return this.headerCheckboxContainer.nativeElement.clientWidth;
+            width += this.headerCheckboxContainer.nativeElement.getBoundingClientRect().width;
+        }
+        if (this.headerDragContainer) {
+            width += this.headerDragContainer.nativeElement.getBoundingClientRect().width;
         }
 
-        return 0;
+        return width;
     }
 
     /**
@@ -2903,7 +2984,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      * @hidden
      */
     get summariesMargin() {
-        return this.rowSelectable ? this.calcRowCheckboxWidth : 0;
+        return this.rowSelectable || this.rowDraggable ? this.headerFixedWidth : 0;
     }
 
     /**
@@ -2952,6 +3033,9 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
 	 * @memberof IgxGridBaseComponent
      */
     public getHeaderGroupWidth(column: IgxColumnComponent): string {
+        if (this.hasColumnLayouts) {
+            return '';
+        }
         const colWidth = column.width;
         const minWidth = this.defaultHeaderGroupMinWidth;
         const isPercentageWidth = colWidth && typeof colWidth === 'string' && colWidth.indexOf('%') !== -1;
@@ -3798,6 +3882,16 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
     get hasColumnGroups(): boolean {
         return this._columnGroups;
     }
+    /**
+     * Returns if the `IgxGridComponent` has column layouts for multi-row layout definition.
+     * ```typescript
+     * const layoutGrid = this.grid.hasColumnLayouts;
+     * ```
+	 * @memberof IgxGridBaseComponent
+     */
+    public get hasColumnLayouts() {
+        return !!this.columnList.some(col => col.columnLayout);
+    }
 
     /**
      * Returns an array of the selected `IgxGridCellComponent`s.
@@ -3812,6 +3906,13 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
                 .reduce((a, b) => a.concat(b), []);
         }
         return [];
+    }
+
+    /**
+     * @hidden
+     */
+    get multiRowLayoutRowSize() {
+        return this._multiRowLayoutRowSize;
     }
 
     /**
@@ -3843,7 +3944,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      * Sets columns defaultWidth property
      */
     protected _derivePossibleWidth() {
-        if (!this._columnWidthSetByUser) {
+        if (!this.columnWidthSetByUser) {
             this._columnWidth = this.getPossibleColumnWidth();
             this.columnList.forEach((column: IgxColumnComponent) => {
                 column.defaultWidth = this._columnWidth;
@@ -3986,8 +4087,22 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
 
         const visibleChildColumns = this.visibleColumns.filter(c => !c.columnGroup);
 
-        const columnsWithSetWidths = visibleChildColumns.filter(c => c.widthSetByUser);
-        const columnsToSize = visibleChildColumns.length - columnsWithSetWidths.length;
+
+        // Column layouts related
+        let visibleCols = [];
+        const columnBlocks = this.visibleColumns.filter(c => c.columnGroup);
+        const colsPerBlock = columnBlocks.map(block => block.getInitialChildColumnSizes(block.children.toArray()));
+        const combinedBlocksSize = colsPerBlock.reduce((acc, item) => acc + item.length, 0);
+        colsPerBlock.forEach(blockCols => visibleCols = visibleCols.concat(blockCols));
+        //
+
+        const columnsWithSetWidths = this.hasColumnLayouts ?
+            visibleCols.filter(c => c.widthSetByUser) :
+            visibleChildColumns.filter(c => c.widthSetByUser);
+
+        const columnsToSize = this.hasColumnLayouts ?
+            combinedBlocksSize - columnsWithSetWidths.length :
+            visibleChildColumns.length - columnsWithSetWidths.length;
 
         const sumExistingWidths = columnsWithSetWidths
             .reduce((prev, curr) => {
@@ -4105,10 +4220,6 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
         this.resetCaches();
         this.calculateGridHeight();
 
-        if (this.showRowCheckboxes) {
-            this.calcRowCheckboxWidth = this.headerCheckboxContainer.nativeElement.getBoundingClientRect().width;
-        }
-
         if (this.rowEditable) {
             this.repositionRowEditingOverlay(this.rowInEditMode);
         }
@@ -4139,9 +4250,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
                 sum += parseInt(col.width, 10);
             }
         }
-        if (this.showRowCheckboxes) {
-            sum += this.calcRowCheckboxWidth;
-        }
+        sum += this.headerFixedWidth;
 
         return sum;
     }
@@ -4263,6 +4372,24 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
      */
     protected initColumns(collection: QueryList<IgxColumnComponent>, cb: Function = null) {
         // XXX: Deprecate index
+        this._columnGroups = this.columnList.some(col => col.columnGroup);
+        if (this.hasColumnLayouts) {
+            // Set overall row layout size
+            this.columnList.forEach((col) => {
+                if (col.columnLayout) {
+                    const layoutSize = col.children ?
+                     col.children.reduce((acc, val) => Math.max(val.rowStart + val.gridRowSpan - 1, acc), 1) :
+                     1;
+                     this._multiRowLayoutRowSize = Math.max(layoutSize, this._multiRowLayoutRowSize);
+                }
+            });
+        }
+        if (this.hasColumnLayouts && this.hasColumnGroups) {
+            // invalid configuration - multi-row and column groups
+            // remove column groups
+            const columnLayoutColumns = this.columnList.filter((col) => col.columnLayout || (col.parent && col.parent.columnLayout));
+            this.columnList.reset(columnLayoutColumns);
+        }
         this._columns = this.columnList.toArray();
         collection.forEach((column: IgxColumnComponent) => {
             column.grid = this;
@@ -4274,7 +4401,6 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
             }
         });
 
-        this._columnGroups = this.columnList.some(col => col.columnGroup);
         this.reinitPinStates();
     }
 
@@ -4664,10 +4790,7 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
             return 0;
         }
 
-        const editModeCell = this.crudService.cell;
-        if (editModeCell) {
-                this.endEdit(false);
-        }
+        this.endEdit(false);
 
         if (!text) {
             this.clearSearch();
@@ -5172,7 +5295,6 @@ export abstract class IgxGridBaseComponent extends DisplayDensityBase implements
             }
         }
     }
-
 }