+++ projects/igniteui-angular/src/lib/grids/grid-common.module.ts
@@ -17,7 +17,7 @@ import { IgxDropDownModule } from '../drop-down/index';
 import { IgxIconModule } from '../icon/index';
 import { IgxInputGroupModule } from '../input-group/input-group.component';
 import { IgxGridCellComponent } from './cell.component';
-import { IgxColumnComponent, IgxColumnGroupComponent } from './column.component';
+import { IgxColumnComponent, IgxColumnGroupComponent, IgxColumnLayoutComponent } from './column.component';
 import { IgxColumnHidingModule } from './column-hiding.component';
 import { IgxGridHeaderComponent } from './grid-header.component';
 import { IgxGridToolbarComponent } from './grid-toolbar.component';
@@ -64,6 +64,7 @@ import { IgxGridPipesModule } from './grid-pipes.module';
 import { IgxGridExcelStyleFilteringModule } from './filtering/excel-style/grid.excel-style-filtering.module';
 import { IgxGridDragSelectDirective } from './drag-select.directive';
 import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
+import { IgxRowDragModule } from './row-drag.directive';
 /**
  * @hidden
  */
@@ -72,6 +73,7 @@ import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
         IgxGridCellComponent,
         IgxColumnComponent,
         IgxColumnGroupComponent,
+        IgxColumnLayoutComponent,
         IgxGridHeaderComponent,
         IgxGridToolbarComponent,
         IgxGridToolbarCustomContentDirective,
@@ -101,12 +103,14 @@ import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
     ],
     entryComponents: [
         IgxColumnComponent,
-        IgxColumnGroupComponent
+        IgxColumnGroupComponent,
+        IgxColumnLayoutComponent
     ],
     exports: [
         IgxGridCellComponent,
         IgxColumnComponent,
         IgxColumnGroupComponent,
+        IgxColumnLayoutComponent,
         IgxGridHeaderComponent,
         IgxGridToolbarComponent,
         IgxGridToolbarCustomContentDirective,
@@ -154,7 +158,8 @@ import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
         IgxFilterModule,
         IgxGridPipesModule,
         IgxGridExcelStyleFilteringModule,
-        IgxFilterCellTemplateDirective
+        IgxFilterCellTemplateDirective,
+        IgxRowDragModule
     ],
     imports: [
         CommonModule,
@@ -181,7 +186,8 @@ import { IgxGridColumnResizerComponent } from './grid-column-resizer.component';
         IgxProgressBarModule,
         IgxFilterModule,
         IgxGridPipesModule,
-        IgxGridExcelStyleFilteringModule
+        IgxGridExcelStyleFilteringModule,
+        IgxRowDragModule
     ],
     providers: [
         IgxGridSelectionService,