+++ projects/igniteui-angular/src/lib/grids/hierarchical-grid/hierarchical-grid.component.ts
@@ -66,6 +66,20 @@ export interface HierarchicalStateRecord {
 })
 export class IgxHierarchicalGridComponent extends IgxHierarchicalGridBaseComponent
     implements IGridDataBindable, AfterViewInit, AfterContentInit, OnInit, OnDestroy {
+
+    /**
+     * @inheritdoc
+     */
+    public get rowDraggable(): boolean {
+        return this.parent ? this.parent.rowDraggable : this._rowDrag;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public set rowDraggable(val: boolean) {
+        this._rowDrag = val;
+    }
     /**
      * Sets the value of the `id` attribute. If not provided it will be automatically generated.
      * ```html
@@ -471,7 +485,7 @@ export class IgxHierarchicalGridComponent extends IgxHierarchicalGridBaseCompone
         return width;
     }
 
-     private getDefaultExpanderWidth(): number {
+    private getDefaultExpanderWidth(): number {
         switch (this.displayDensity) {
             case DisplayDensity.cosy:
                 return 57;
@@ -557,6 +571,19 @@ export class IgxHierarchicalGridComponent extends IgxHierarchicalGridBaseCompone
             currGrid = currGrid.parent;
         }
         return currGrid;
+    }
+
+    /**
+     * @hidden
+    */
+    protected initColumns(collection: QueryList<IgxColumnComponent>, cb: Function = null) {
+        if (this.hasColumnLayouts) {
+            // invalid configuration - hierarchical grid should not allow column layouts
+            // remove column layouts
+            const nonColumnLayoutColumns = this.columnList.filter((col) => !col.columnLayout && !(col.parent && col.parent.columnLayout));
+            this.columnList.reset(nonColumnLayoutColumns);
+        }
+        super.initColumns(collection, cb);
     }
 
     /**